Create getRandomGif
// This code assumes that you have a list of GIF files in a folder and their names are stored 

in the 'gifFiles' array. The 'selectRandomGif' function selects a random GIF file from the list 

using the 'Math.random' function and returns the selected GIF file. The 'uploadGifAsSprite' 

function is responsible for uploading the selected GIF file as a sprite in Scratch. It creates a 

new image element, sets its 'src' attribute to the GIF file, and waits for the image to load. 

// Once the image is loaded, it gets the canvas element from the Scratch editor, gets the 

context of the canvas, and draws the image on the canvas. It then gets the image data from 

the canvas and creates a new sprite in the Scratch editor with the image data.